requirements.txt
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.4.0
plantcv==3.13.0
pyzbar>=0.1.8
